# doc/apidoc/apidoc.json
{
  "swagger": "2.0",
  "info": {
    "description": "风早早和艾小小的express后端项目的api",
    "version": "1.0.0",
    "title": "express",
    "contact": {
      "email": "2323504502@qq.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8002",
  "basePath": "/apidoc",
...
  "schemes": [
    "http",
    "https"
  ],
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
  "paths": {
...
    "/user/v1/getYuan": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "getYuan",
        "description": "获取心愿",
        "operationId": "getYuan",
        "parameters": [
          {
            "$ref": "#/parameters/userAgent"
          },
          {
            "name": "status",
            "in": "query",
            "description": "心愿状态1进行中 2待审核 3 完成 ",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageNo",
            "in": "query",
            "description": "pageNo",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "reason",
            "in": "query",
            "description": "搜索方式1关键字搜索 2首页获取 3自己获取",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "搜索关键字",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "0": {
            "description": "fail operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/user/v1/getUserInfo": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "get userInfo",
        "description": "获取微信用户信息",
        "operationId": "getUserInfo",
        "parameters": [
          {
            "$ref": "#/parameters/userAgent"
          }
        ],
        "responses": {
          "0": {
            "description": "fail operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/admin/v1/getUserInfo": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "userInfo",
        "description": "获取用户信息",
        "operationId": "userInfo",
        "parameters": [
          {
            "$ref": "#/parameters/userAgent"
          }
        ],
        "responses": {
          "0": {
            "description": "fail operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/admin/v1/getYuan": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "getYuan",
        "description": "获取所有心愿",
        "operationId": "getYuan",
        "parameters": [
          {
            "$ref": "#/parameters/userAgent"
          }
        ],
        "responses": {
          "0": {
            "description": "fail operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
    "/admin/v1/getUser": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "getUser",
        "description": "获取所有用户",
        "operationId": "getUser",
        "parameters": [
          {
            "$ref": "#/parameters/userAgent"
          }
        ],
        "responses": {
          "0": {
            "description": "fail operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    }
  },
  "parameters": {
    "userAgent": {
      "name": "Authorization",
      "type": "string",
      "in": "header",
      "required": true,
      "default": "Bearer "
    }
  },
...
}